Add files via upload
Project-2

After the initial analysis, we decided to use models like Logistic Regression, Random Forest, Decision Trees, kNN, & Naïve-Bayes Classifier for this analysis. We split the data into train & test & built a model using each of these classifiers. We used K-fold Cross-validation to evaluate the quality of the predictive models by partitioning the original data into a training set to train the model, and a test set to evaluate it.